+++ ichnaea/api/locate/query.py
@@ -253,6 +253,16 @@ class Query(object):
         # return the best possible (smallest) accuracy
         return min(accuracies)
 
+    def result_status(self, result):
+        """
+        Returns either hit or miss, depending on whether the result
+        matched the expected query accuracy.
+        """
+        if result.data_accuracy <= self.expected_accuracy:
+            # equal or better / smaller accuracy
+            return 'hit'
+        return 'miss'
+
     def internal_query(self):
         """Returns a dictionary of this query in our internal format."""
         result = {}
@@ -280,18 +290,14 @@ class Query(object):
     def collect_metrics(self):
         """Should detailed metrics be collected for this query?"""
         allowed = bool(self.api_key and self.api_key.log and self.api_type)
-        # don't report stats if there is really no data at all
-        # in the query
+        # don't report stats if there is no data at all in the query
         possible_result = bool(self.expected_accuracy != DataAccuracy.none)
-        return allowed and possible_result
+        return (allowed and possible_result)
 
     def _emit_country_stat(self, metric, extra_tags):
-        # Emit a country specific stat
-        if not self.country:
+        country = self.country
+        if not country:
             country = 'none'
-        else:
-            # TODO emit actual country based stats
-            country = 'xx'
 
         metric = '%s.%s' % (self.api_type, metric)
         tags = [
@@ -322,16 +328,7 @@ class Query(object):
         if not self.collect_metrics():
             return
 
-        if self.api_type == 'country':
-            # TODO: enable result stats for country API
-            return
-
-        if result.data_accuracy <= self.expected_accuracy:
-            # equal or better / smaller accuracy
-            status = 'hit'
-        else:
-            status = 'miss'
-
+        status = self.result_status(result)
         tags = [
             'accuracy:%s' % self.expected_accuracy.name,
             'status:%s' % status,
@@ -341,15 +338,11 @@ class Query(object):
         self._emit_country_stat('result', tags)
 
     def emit_source_stats(self, source, result):
-        """Emit stats about how well the result satisfied the query."""
+        """Emit stats about how well the source satisfied the query."""
         if not self.collect_metrics():
             return
 
-        if result.data_accuracy <= self.expected_accuracy:
-            status = 'hit'
-        else:
-            status = 'miss'
-
+        status = self.result_status(result)
         tags = [
             'source:%s' % source.name,
             'accuracy:%s' % self.expected_accuracy.name,